restructured and simplified the test file
