+++ cypress/pageObject/Patients/PatientEncounter.ts
@@ -222,118 +222,8 @@ export class PatientEncounter {
     return this;
   }
 
-  /*** ENCOUNTER_NOTES ***/
-
-  // **1️⃣ Intercept Setup**
-  interceptSendMessageRequest() {
-    cy.intercept("POST", "/api/v1/patient/*/thread/*/note/").as("sendMessage");
-    return this;
-  }
-
-  // **2️⃣ Encounter Handling**
-  openActiveEncounterAndSaveId() {
-    cy.get("[data-cy^='encounter-card-']")
-      .not((_, el) =>
-        this.inactiveEncounterStatus.includes(el.getAttribute("data-status")),
-      )
-      .first()
-      .then(($el) => {
-        const encounterId = $el.attr("data-cy")?.replace("encounter-card-", "");
-        if (!encounterId) throw new Error("Encounter ID is undefined!");
-
-        Cypress.env("encounterId", encounterId);
-        this.openEncounterById(encounterId);
-      });
-
-    return this;
-  }
-
-  openEncounterById(encounterId: string) {
-    cy.verifyAndClickElement(
-      `[data-cy="encounter-card-${encounterId}"]`,
-      "View Details",
-    );
-    return this;
-  }
-
   openEncounterNotesTab() {
-    cy.verifyAndClickElement('[data-cy="encounter-notes-tab"]', "Notes");
-    return this;
-  }
-
-  // **3️⃣ Thread Management**
-  clickNewThreadButton() {
-    cy.verifyAndClickElement('[data-cy="new-thread-button"]', "New");
-    return this;
-  }
-
-  typeThreadTitle(title: string) {
-    cy.typeIntoField('[data-cy="new-thread-title-input"]', title);
-    return this;
-  }
-
-  clickCreateThreadButton() {
-    cy.verifyAndClickElement('[data-cy="create-thread-button"]', "Create");
-    return this;
-  }
-
-  // **4️⃣ Message Handling**
-  typeMessage(message: string) {
-    cy.get('[data-cy="encounter-notes-chat-message-input"]')
-      .should("be.visible")
-      .clear()
-      .type(message);
-    return this;
-  }
-
-  sendMessage(message: string) {
-    if (!message.trim()) return;
-
-    this.typeMessage(message);
-    cy.get('[data-cy="send-chat-message-button"]')
-      .should("be.visible")
-      .should("not.be.disabled")
-      .click();
-
-    cy.wait("@sendMessage").its("response.statusCode").should("eq", 200);
-  }
-
-  addNewChatMessages(messages: string[]) {
-    this.interceptSendMessageRequest();
-
-    messages.forEach((message) => {
-      this.sendMessage(message);
-    });
-
-    return this;
-  }
-
-  verifyMessagesInChat(messages: string[]) {
-    cy.verifyContentPresence('[data-cy="chat-messages"]', messages);
-    return this;
-  }
-
-  verifyMessagesNotExistInChat(messages: string[]) {
-    messages.forEach((message) => {
-      cy.get('[data-cy="chat-messages"]').contains(message).should("not.exist");
-    });
-    return this;
-  }
-
-  // **5️⃣ Thread Switching**
-  changeThread(title: string) {
-    cy.get('[data-cy="thread-title"]')
-      .should("be.visible")
-      .contains(title)
-      .click();
-    return this;
-  }
-
-  // **6️⃣ Logout**
-  logout() {
-    cy.visit("/");
-    cy.get('[data-cy="user-dashboard-dropdown"]').click();
-    cy.verifyAndClickElement('[data-cy="sign-out-button"]', "Sign out");
+    cy.verifyAndClickElement('[data-cy="tab-notes"]', "Notes");
     return this;
   }
 }
